Added DTO models - WorkshopWithRequiredPropertiesDto, WorkshopWithDescriptionDto, WorkshopWithContactsDto and WorkshopWithTeachersDto.
