+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/IncompletedWorkshops/WorkshopWithContactsDto.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Util.JsonTools;
+
+namespace OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
+public class WorkshopWithContactsDto : WorkshopWithDescriptionDto
+{
+    [Required]
+    public long AddressId { get; set; }
+
+    [Required]
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    public AddressDto Address { get; set; }
+}
