+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/IncompletedWorkshops/WorkshopWithDescriptionDto.cs
@@ -1,0 +1,30 @@
+﻿using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Util.CustomValidation;
+using OutOfSchool.BusinessLogic.Util.JsonTools;
+
+namespace OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
+public class WorkshopWithDescriptionDto : WorkshopWithRequiredPropertiesDto
+{
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    [CollectionNotEmpty(ErrorMessage = "At least one description is required")]
+    public IEnumerable<WorkshopDescriptionItemDto> WorkshopDescriptionItems { get; set; }
+
+    public bool WithDisabilityOptions { get; set; } = default;
+
+    [MaxLength(200)]
+    public string DisabilityOptionsDesc { get; set; } = string.Empty;
+
+    public Guid? InstitutionId { get; set; }
+
+    public string Institution { get; set; }
+
+    public Guid? InstitutionHierarchyId { get; set; }
+
+    public string InstitutionHierarchy { get; set; }
+
+    public List<long> DirectionIds { get; set; }
+
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    public IEnumerable<string> Keywords { get; set; } = default;
+}
