+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/IncompletedWorkshops/WorkshopWithTeachersDto.cs
@@ -1,0 +1,5 @@
+﻿namespace OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
+public class WorkshopWithTeachersDto : WorkshopWithContactsDto
+{
+    public List<TeacherDTO> Teachers { get; set; }
+}
