+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/WorkshopMementoController.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
 using OutOfSchool.BusinessLogic.Services.Memento.Interfaces;
-using OutOfSchool.BusinessLogic.Services.Memento.Models;
 
 namespace OutOfSchool.WebApi.Controllers.V1;
 
-/// <summary>Controller with CRUD operations for IncomplitedWorkshopDto entity.</summary>
+/// <summary>Controller with CRUD operations for WorkshopWithRequiredPropertiesDto entity.</summary>
 [ApiController]
 [ApiVersion("1.0")]
 [Route("api/v{version:apiVersion}/[controller]/[action]")]
-public class WorkshopMementoController(IMementoService<IncompletedWorkshopDto> mementoService)
-    : BaseMementoController<IncompletedWorkshopDto>(mementoService)
+public class WorkshopMementoController(IMementoService<WorkshopWithRequiredPropertiesDto> mementoService)
+    : BaseMementoController<WorkshopWithRequiredPropertiesDto>(mementoService)
 {
 }
